Added ability to seach from commit hash from version
Now if you run the project with '-v (version)', the corresponding commit hash to the version you input will be returned. Be aware this a work in progress and is NOT AT ALL fast. Searching for the oldest recorded version on Dolphin's website took slightly less than 6.5 minutes for my computer to return the corresponding hash. I do have plans to potentially reduce the time.